requirements.txt
requests
polars[all]
sqlalchemy
faker
pandas
pyyaml
pymysql
pyodbc
pymssql